make color primary key in categoryEntity, add getFilters in categoryRepo, add categoryRepo in di, add logic for filters in SitesViewModel

+++ features/main/src/main/java/bes/max/features/main/domain/models/FilterModel.kt
@@ -5,5 +5,5 @@ import androidx.compose.ui.graphics.Color
 data class FilterModel(
     val name: String? = null,
     val color: Color,
-    val filterAction: () -> Unit,
+    val filterAction: (Int) -> Unit,
 )
